refactor (src/*): refactored http and ftp logic
* create a core module and a download module that extends the core
