+++ src/lib.rs
@@ -5,7 +5,7 @@ extern crate reqwest;
 extern crate indicatif;
 extern crate num_cpus;
 
-pub mod http;
-pub mod rftp;
+pub mod download;
 pub mod utils;
 pub mod bar;
+pub mod core;
